requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-extensions==3.2.3
django-ninja==1.1.0
pydantic==2.7.4
pydantic_core==2.18.4
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
